Implement authrization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -64,6 +64,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::allows('update-question', $question)) {
+            abort(403, "Access denied");
+
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -76,6 +80,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::allows('update-question', $question)) {
+            abort(403, "Access denied");
+
+        }
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', "Your question has benn updated.");
@@ -89,6 +97,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::allows('delete-question', $question)) {
+            abort(403, "Access denied");
+
+        }
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been delete");
